+++ tests/integration/test_credit_card.py
@@ -16,7 +16,7 @@ class TestCreditCard(unittest.TestCase):
 
         self.assertTrue(result.is_success)
         credit_card = result.credit_card
-        self.assertTrue(re.search("\A\w{4,5}\Z", credit_card.token) is not None)
+        self.assertTrue(re.search("\A\w{4,}\Z", credit_card.token) is not None)
         self.assertEquals("411111", credit_card.bin)
         self.assertEquals("1111", credit_card.last_4)
         self.assertEquals("05", credit_card.expiration_month)
@@ -478,7 +478,7 @@ class TestCreditCard(unittest.TestCase):
 
         self.assertTrue(result.is_success)
         credit_card = result.credit_card
-        self.assertTrue(re.search("\A\w{4,5}\Z", credit_card.token) is not None)
+        self.assertTrue(re.search("\A\w{4,}\Z", credit_card.token) is not None)
         self.assertEquals("510510", credit_card.bin)
         self.assertEquals("5100", credit_card.last_4)
         self.assertEquals("06", credit_card.expiration_month)
@@ -694,7 +694,7 @@ class TestCreditCard(unittest.TestCase):
         }).credit_card
 
         found_credit_card = CreditCard.find(credit_card.token)
-        self.assertTrue(re.search("\A\w{4,5}\Z", credit_card.token) is not None)
+        self.assertTrue(re.search("\A\w{4,}\Z", credit_card.token) is not None)
         self.assertEquals("411111", credit_card.bin)
         self.assertEquals("1111", credit_card.last_4)
         self.assertEquals("05", credit_card.expiration_month)
@@ -721,12 +721,9 @@ class TestCreditCard(unittest.TestCase):
         self.assertEquals(Decimal("1.00"), found_credit_card.subscriptions[0].price)
         self.assertEquals(credit_card.token, found_credit_card.subscriptions[0].payment_method_token)
 
+    @raises_with_regexp(NotFoundError, "payment method with token 'bad_token' not found")
     def test_find_with_invalid_token(self):
-        try:
-            CreditCard.find("bad_token")
-            self.assertTrue(False)
-        except Exception as e:
-            self.assertEquals("payment method with token 'bad_token' not found", str(e))
+        CreditCard.find("bad_token")
 
     def test_from_nonce_with_unlocked_nonce(self):
         config = Configuration.instantiate()
@@ -756,6 +753,7 @@ class TestCreditCard(unittest.TestCase):
         customer = Customer.find(customer.id)
         self.assertEquals(customer.credit_cards[0].token, card.token)
 
+    @raises_with_regexp(NotFoundError, "payment method with nonce .* or not found")
     def test_from_nonce_with_unlocked_nonce_pointing_to_shared_card(self):
         config = Configuration.instantiate()
 
@@ -778,12 +776,9 @@ class TestCreditCard(unittest.TestCase):
         self.assertEqual(status_code, 201)
         nonce = json.loads(response)["creditCards"][0]["nonce"]
 
-        try:
-            CreditCard.from_nonce(nonce)
-            self.assertTrue(False)
-        except Exception as e:
-            self.assertIn("not found", str(e))
+        CreditCard.from_nonce(nonce)
 
+    @raises_with_regexp(NotFoundError, ".* consumed .*")
     def test_from_nonce_with_consumed_nonce(self):
         config = Configuration.instantiate()
         customer = Customer.create().customer
@@ -809,11 +804,7 @@ class TestCreditCard(unittest.TestCase):
         nonce = json.loads(response)["creditCards"][0]["nonce"]
 
         CreditCard.from_nonce(nonce)
-        try:
-            CreditCard.from_nonce(nonce)
-            self.assertTrue(False)
-        except Exception as e:
-            self.assertIn("consumed", str(e))
+        CreditCard.from_nonce(nonce)
 
     def test_create_from_transparent_redirect(self):
         customer = Customer.create().customer