+++ tests/integration/test_oauth.py
@@ -82,12 +82,33 @@ class TestOAuthGateway(unittest.TestCase):
         self.assertIsNotNone(credentials.expires_at)
         self.assertEquals("bearer", credentials.token_type)
 
+    def test_revoke_access_token(self):
+        code = TestHelper.create_grant(self.gateway, {
+            "merchant_public_id": "integration_merchant_id",
+            "scope": "read_write"
+        })
+
+        access_token = self.gateway.oauth.create_token_from_code({
+            "code": code,
+            "scope": "read_write"
+        }).credentials.access_token
+
+        result = self.gateway.oauth.revoke_access_token(access_token)
+
+        self.assertTrue(result.is_success)
+
+        with self.assertRaises(AuthenticationError) as error:
+            gateway = BraintreeGateway(access_token = access_token)
+
+            gateway.customer.create()
+
     def test_connect_url(self):
         connect_url = self.gateway.oauth.connect_url({
              "merchant_id": "integration_merchant_id",
              "redirect_uri": "http://bar.example.com",
              "scope": "read_write",
              "state": "baz_state",
+             "landing_page": "login",
              "user": {
                "country": "USA",
                "email": "foo@example.com",
@@ -130,6 +151,7 @@ class TestOAuthGateway(unittest.TestCase):
         self.assertEqual(params["redirect_uri"], ["http://bar.example.com"])
         self.assertEqual(params["scope"], ["read_write"])
         self.assertEqual(params["state"], ["baz_state"])
+        self.assertEqual(params["landing_page"], ["login"])
 
         self.assertEqual(params["user[country]"], ["USA"])
         self.assertEqual(params["business[name]"], ["14 Ladders"])