+++ tests/unit/test_http.py
@@ -4,12 +4,13 @@ from tests.test_helper import *
 from braintree.attribute_getter import AttributeGetter
 
 class TestHttp(unittest.TestCase):
+    @raises(UpgradeRequiredError)
     def test_raise_exception_from_status_for_upgrade_required(self):
-        try:
-            Http.raise_exception_from_status(426)
-            self.assertTrue(False)
-        except UpgradeRequiredError:
-            pass
+        Http.raise_exception_from_status(426)
+
+    @raises(TooManyRequestsError)
+    def test_raise_exception_from_too_many_requests(self):
+        Http.raise_exception_from_status(429)
 
     def test_backtrace_preserved_when_not_wrapping_exceptions(self):
         class Error(Exception):
@@ -27,7 +28,7 @@ class TestHttp(unittest.TestCase):
                 "wrap_http_exceptions": False})
 
         try:
-            Http(config, "a")._Http__http_do("a", "b")
+            Http(config, "fake_environment").post("/example/path/to/reach")
         except Error as e:
             _, _, tb = sys.exc_info()
             self.assertEqual('raise_error', traceback.extract_tb(tb)[-1][2])